+++ lib/main.dart
@@ -63,8 +63,9 @@ data
 
     final typeString = argResults[ConstArgOptionFlag.typeOption].toString();
     final typeRun = TypeRun.fromValue(typeString, fallback: TypeRun.none);
-    logger..info('Type of generator used: `${typeRun.value}`')
-    ..progress('load');
+    logger
+      ..info('Type of generator used: `${typeRun.value}`')
+      ..progress('load');
     switch (typeRun) {
       case TypeRun.enumDefault:
         await runEnumDefault(path: path, logger: logger);